+++ src/main/java/net/bqc/aoc/year2023/Day04.java
@@ -8,11 +8,11 @@ import java.util.List;
 public class Day04 extends Solution {
 
     public static class ScratchCard {
-        int gameId;
-        int[] slots = new int[100];
-        int winningPoints = 0;
-        int instances = 1;
-        int count = 0;
+        long gameId;
+        long[] slots = new long[100];
+        long winningPoints = 0;
+        long instances = 1;
+        long count = 0;
 
         @Override
         public String toString() {
@@ -26,16 +26,16 @@ public class Day04 extends Solution {
     }
 
     @Override
-    public int solve(PART_NUMBER part, List<String> inputLines) {
+    public long solve(PART_NUMBER part, List<String> inputLines) {
         this.pup = part;
 
         if (this.pup == PART_NUMBER.ONE) {
-            return inputLines.stream().mapToInt(x -> parseCardPoint(x).winningPoints).sum();
+            return inputLines.stream().mapToLong(x -> parseCardPoint(x).winningPoints).sum();
         }
         else {
             List<ScratchCard> cards = inputLines.stream().map(this::parseCardPoint).toList();
             for (int i = 0; i < cards.size(); i++) {
-                int matchedCount = cards.get(i).count;
+                long matchedCount = cards.get(i).count;
                 if (matchedCount >= 1) {
                     for (int j = 1; j <= matchedCount && i + j < cards.size(); j++) {
                         cards.get(i + j).instances += cards.get(i).instances;
@@ -43,7 +43,7 @@ public class Day04 extends Solution {
                 }
             }
 
-            return cards.stream().mapToInt(x -> x.instances).sum();
+            return cards.stream().mapToLong(x -> x.instances).sum();
         }
     }
 
@@ -51,7 +51,7 @@ public class Day04 extends Solution {
         ScratchCard card = new ScratchCard();
 
         String[] parts = cardInfo.split(":");
-        card.gameId = Integer.parseInt(parts[0].split("\s+")[1]);
+        card.gameId = Long.parseLong(parts[0].split("\s+")[1]);
 
         List<Integer> winningNumbers = Arrays.stream(parts[1].split("\\|")[0].trim().split("\s+"))
             .map(Integer::parseInt).toList();
@@ -61,10 +61,10 @@ public class Day04 extends Solution {
 
         winningNumbers.forEach(x -> card.slots[x] = 1);
 
-        int count = (int) guessingNumbers.stream().filter(x -> card.slots[x] > 0).count();
+        long count = (long) guessingNumbers.stream().filter(x -> card.slots[x] > 0).count();
 
         card.count = count;
-        card.winningPoints = count <= 1 ? count : (int)Math.pow(2, count - 1);
+        card.winningPoints = count <= 1 ? count : (long)Math.pow(2, count - 1);
         return card;
     }
 }
